# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plataforma.dev API
  description: |
    Complete API documentation for Plataforma.dev - the AI-first enterprise platform.
    
    This API provides comprehensive endpoints for:
    - User authentication and authorization
    - Module management and deployment
    - Database operations with real-time updates
    - File storage and management
    - Real-time communication via WebSockets
    
    ## Authentication
    
    The API uses JWT (JSON Web Tokens) for authentication. Include the token in the Authorization header:
    
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    API requests are rate-limited to ensure fair usage:
    - **General API**: 1000 requests per hour per IP
    - **Authentication**: 10 attempts per 15 minutes per IP
    - **File uploads**: 100 requests per hour per user
    
    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Request limit per time window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Time when the rate limit resets
    
    ## Error Handling
    
    The API uses conventional HTTP response codes and returns detailed error information:
    
    ```json
    {
      "success": false,
      "error": {
        "code": "VALIDATION_FAILED",
        "message": "Request validation failed",
        "details": {
          "field": "email",
          "reason": "Invalid email format"
        }
      },
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: Retrieve the profile of the authenticated user
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Users
      summary: Update user profile
      description: Update the current user's profile information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: User ID
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve a specific user's information
      responses:
        '200':
          description: User information retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Users
      summary: Update user (Admin only)
      description: Update a user's information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Users
      summary: Delete user (Admin only)
      description: Delete a user account
      responses:
        '204':
          description: User deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Module Management Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          example: John Doe
        avatar:
          type: string
          format: uri
          nullable: true
        roles:
          type: array
          items:
            type: string
          example: [user, manager]
        permissions:
          type: array
          items:
            type: string
        preferences:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
          nullable: true
...
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        avatar:
          type: string
          format: uri
        preferences:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_FAILED
            message:
              type: string
              example: Request validation failed
            details:
              type: object
              additionalProperties: true
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_FAILED
              message: Request validation failed
              details:
                field: email
                reason: Invalid email format
            timestamp: 2024-08-26T10:30:00Z
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
            timestamp: 2024-08-26T10:30:00Z
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Insufficient permissions
            timestamp: 2024-08-26T10:30:00Z
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
            timestamp: 2024-08-26T10:30:00Z
...